+++ src/ray/raylet_client/raylet_client.h
@@ -317,7 +317,7 @@ class RayletClient : public RayletClientInterface {
                                  bool mark_worker_blocked,
                                  const TaskID &current_task_id,
                                  const std::vector<unsigned long> &batch_virt_address,
-                                 const std::vector<int64_t> &batch_object_size);
+                                 const std::vector<int> &batch_object_size);
 
   /// Notify the raylet that this client (worker) is no longer blocked.
   ///
